requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.19.0
openpyxl==3.0.4
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
